requirements.txt
yfinance==0.2.36
pandas==2.1.4
numpy==1.26.4
pandas-ta==0.3.14b
scikit-learn==1.4.0
anthropic==0.23.1
streamlit==1.32.0
plotly==5.19.0
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
joblib==1.3.2
pytz==2024.1
ta==0.11.0
nsepy==0.8
